Split out ESPColorCorrection into a separate file
